requirements.txt
click==8.0.1
Flask==2.0.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
protobuf==3.17.3
six==1.16.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
